added roll generation to server so rolls appear to be the same for all connected clients/users
