+++ src/components/ChatPanel/Chat/MessageList/Message/RollMessage.js
@@ -1,25 +1,29 @@
 import React from 'react';
 import './Message.css';
-import { getTotalDamage, rollAttack, rollDamage } from "./roll";
 
-const RollMessage = ({message: {name, creatureName, action, createdAt}, clientName}) => {
-	const [rolls, setRolls] = React.useState({att:0, dmg:[], isCrit:false});
+const RollMessage = ({message: {name, creatureName, action, rolls, createdAt}, clientName}) => {
+	// const [rolls, setRolls] = React.useState({att:0, dmg:[], isCrit:false});
   const isCurrentUser = name === clientName.trim() ? true : false;
 	
-	const getRolls = () => {
-		let isCrit = false;
-		const att = rollAttack();
-		if(att === 20 || att === '20') {
-			isCrit = true;
-		}
-		const dmg = rollDamage(action.damage_dice, action.damage_bonus, isCrit);
-		// console.log("{att, dmg, isCrit}: ",{att: att, dmg: dmg, isCrit: isCrit})
-		return { att: att, dmg: dmg, isCrit: isCrit }; // { number, array, boolean } 
-	}
+	// const getRolls = () => {
+	// 	let isCrit = false;
+	// 	const att = rollAttack();
+	// 	if(att === 20 || att === '20') {
+	// 		isCrit = true;
+	// 	}
+	// 	const dmg = rollDamage(action.damage_dice, action.damage_bonus, isCrit);
+	// 	// console.log("{att, dmg, isCrit}: ",{att: att, dmg: dmg, isCrit: isCrit})
+	// 	return { att: att, dmg: dmg, isCrit: isCrit }; // { number, array, boolean } 
+	// }
+	const getTotalDamage = (array) => {
+		let total = 0;
+		array.map(el => total += el);
+		return total; // number
+	};
 	
-	React.useEffect(() => {
-		setRolls(getRolls());
-	}, []);
+	// React.useEffect(() => {
+	// 	setRolls(getRolls());
+	// }, []);
 
 	return (
 		<div className={`message-container mt1 pl2 pr2 ${isCurrentUser ? "justify-end" : "justify-start"}`}>
