+++ src/server/index.js
@@ -1,4 +1,3 @@
-// const path = require("path");
 const http = require("http");
 const express = require("express");
 const cors = require("cors");
@@ -17,13 +16,11 @@ const io = socketio(server);
 const port = process.env.PORT || 5016;
 // https://whispering-brook-74854.herokuapp.com/
 
-// let buggySendRollMessage = 0;
-// let buggySendMessage = 0;
 io.on("connection", (socket) => {
   console.log("New WebSocket connection: " + socket.id);
   
   socket.on("join", ({name, room}, callback) => {
-    console.log("socket.on('join'), from SERVER");
+    // console.log("socket.on('join'), from SERVER");
     const { error, user } = addUser({ id: socket.id, name, room});
 
     if(error) return callback(error);
@@ -31,8 +28,6 @@ io.on("connection", (socket) => {
     socket.join(user.room);
 
     socket.emit("message", conjureChatMessage("Innkeeper",`${user.name}, Welcome!`));
-    let test = conjureChatMessage(user.name, "TEST");
-    console.log(test, typeof test);
     socket.broadcast.to(user.room).emit("message", conjureChatMessage("Innkeeper", `${user.name} has joined!`));
 
     io.to(user.room).emit("roomData", { room: user.room, users: getUsersInRoom(user.room) });
@@ -42,19 +37,15 @@ io.on("connection", (socket) => {
 
   socket.on("sendMessage", (text, callback) => {
     const user = getUser(socket.id);
-    io.to(user.room).emit("message", conjureChatMessage(user.name, text));
-    
-    console.log("sendMessage fired, from SERVER: ", conjureChatMessage(user.name, text));
-    // console.log(++buggySendMessage); // expect +1 every message
+		io.to(user.room).emit("message", conjureChatMessage(user.name, text));
+		
     callback();
   });
 
   socket.on("sendRollMessage", ({creatureName, action}, callback) => {
     const user = getUser(socket.id);
-    io.to(user.room).emit("message", conjureRollMessage(user.name, creatureName, action));
-    
-    console.log("SendRollMessage fired, from SERVER: ", conjureRollMessage(user.name, creatureName, action));
-    // console.log(++buggySendRollMessage); // expect +2 every roll if bug is coming from server
+		io.to(user.room).emit("message", conjureRollMessage(user.name, creatureName, action));
+		
     callback();
   });
 
@@ -71,4 +62,3 @@ io.on("connection", (socket) => {
 });
 
 server.listen(port, () => console.log(`Server is running on port *: ${port}`));
-// console.log(`socket.io server is up on port *: ${port}`);
